Change testDecodeInt() case for reverse behavior
This commit replaces an assertion that was dropped in the previous commit.

Originally, the assertion was that `decode_int("00101", ...)` yields `101`.
With the change in the previous commit to decode the *reverse* of the input,
the assertion was changed to be that `decode_int_r("00101", ...)` yields
`10100`. However, this is not the same assertion as we are no longer asserting
that leading zeroes are not affecting the output of the method. This commit
effectively reverts the assertion to be that `decode_int_r("10100", ...)`
(i.e. the reverse of "00101") yields `101`.